feat: fix casting of long to BigInteger for EventTriggerBriefInfo. Change version to 5.1.30-SNAPSHOT.

+++ src/main/java/org/qubership/automation/itf/core/hibernate/spring/managers/executor/SituationEventTriggerObjectManager.java
@@ -76,7 +76,7 @@ public class SituationEventTriggerObjectManager extends EventTriggerObjectManage
                     continue;
                 }
                 EventTriggerBriefInfo briefInfo = new EventTriggerBriefInfo(
-                        (BigInteger) entry[0],
+                        BigInteger.valueOf((Long) entry[0]),
                         SituationEventTrigger.TYPE);
                 if ("ACTIVE".equals(entry[1].toString())) {
                     triggersToReactivate.add(briefInfo);
